+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased] 3.6.2.1
 ### Fixed
 - Fix migration issue when no matrix or SuperTable blocks exist (Fixes #182)
+- Fix issue with Google trying to set legacy parts that aren’t supported (Fixes #183)
 
 ## 3.6.2 - 2019-07-23
 ### Added